+++ cmd/start.go
@@ -1,12 +1,45 @@
 package cmd
 
 import (
+	"fmt"
+	"os"
+	"sort"
 	"strings"
 
 	"github.com/spf13/afero"
 	"github.com/spf13/cobra"
 	"github.com/supabase/cli/internal/start"
+	"github.com/supabase/cli/internal/utils"
 )
+
+func validateExcludedContainers(excludedContainers []string) {
+	// Validate excluded containers
+	validContainers := start.ExcludableContainers()
+	var invalidContainers []string
+
+	for _, e := range excludedContainers {
+		found := false
+		for _, v := range validContainers {
+			if strings.EqualFold(e, v) {
+				found = true
+				break
+			}
+		}
+		if !found {
+			invalidContainers = append(invalidContainers, e)
+		}
+	}
+
+	if len(invalidContainers) > 0 {
+		// Sort the names list so it's easier to visually spot the one you looking for
+		sort.Strings(validContainers)
+		warning := fmt.Sprintf("%s The following container names are not valid to exclude: %s\nValid containers to exclude are: %s\n",
+			utils.Yellow("Warning:"),
+			utils.Aqua(strings.Join(invalidContainers, ", ")),
+			utils.Aqua(strings.Join(validContainers, ", ")))
+		fmt.Fprint(os.Stderr, warning)
+	}
+}
 
 var (
 	allowedContainers  = start.ExcludableContainers()
@@ -19,6 +52,7 @@ var (
 		Use:     "start",
 		Short:   "Start containers for Supabase local development",
 		RunE: func(cmd *cobra.Command, args []string) error {
+			validateExcludedContainers(excludedContainers)
 			return start.Run(cmd.Context(), afero.NewOsFs(), excludedContainers, ignoreHealthCheck)
 		},
 	}
