+++ internal/start/start.go
@@ -10,7 +10,6 @@ import (
 	"os"
 	"path"
 	"path/filepath"
-	"sort"
 	"strconv"
 	"strings"
 	"text/template"
@@ -44,39 +43,9 @@ func suggestUpdateCmd(serviceImages map[string]string) string {
 	return cmd
 }
 
-func validateExcludedContainers(excludedContainers []string) {
-	// Validate excluded containers
-	validContainers := ExcludableContainers()
-	var invalidContainers []string
-
-	for _, e := range excludedContainers {
-		found := false
-		for _, v := range validContainers {
-			if strings.EqualFold(e, v) {
-				found = true
-				break
-			}
-		}
-		if !found {
-			invalidContainers = append(invalidContainers, e)
-		}
-	}
-
-	if len(invalidContainers) > 0 {
-		// Sort the names list so it's easier to visually spot the one you looking for
-		sort.Strings(validContainers)
-		warning := fmt.Sprintf("%s The following container names are not valid to exclude: %s\nValid containers to exclude are: %s\n",
-			utils.Yellow("Warning:"),
-			utils.Aqua(strings.Join(invalidContainers, ", ")),
-			utils.Aqua(strings.Join(validContainers, ", ")))
-		fmt.Fprint(os.Stderr, warning)
-	}
-}
-
 func Run(ctx context.Context, fsys afero.Fs, excludedContainers []string, ignoreHealthCheck bool) error {
 	// Sanity checks.
 	{
-		validateExcludedContainers(excludedContainers)
 		if err := utils.LoadConfigFS(fsys); err != nil {
 			return err
 		}
@@ -859,7 +828,7 @@ EOF
 	}
 
 	// Start Storage.
-	if isStorageEnabled && !isContainerExcluded(utils.Config.Storage.Image, excluded) {
+	if isStorageEnabled {
 		dockerStoragePath := "/mnt"
 		if _, err := utils.DockerStart(
 			ctx,
